QA Updates - Round 1 (#21)
* Menu: put sharing first

* Update links

* Rename Quickstart guide, update references

* Remove thumbnail on top of blogs

* Update quickstart guide link

* Create Docs page

* Create subpages under Community

* Tweak menus

* Update sharing page hero images

* Re-use video list on Learn page

* Learn page: list tutorials

* Add contributors from Github

* List meetings (using fake data for now)

* Add social and org logo tiles to community page

- Created shared Tiles component
- Moved social and org logos to tiles

This did have a compromise of reverting the hover style. I think this is
more consistent with other examples of hovering over link blocks but I
would be happy to discuss a different approach. I just don't think the
scale animation works on the light variation and I wanted to simplify
the styles in the process.

* Add logo grid

* Open LF link in new tab