Use rehype plugins for slugs/syntax highlighting
It appears that something is getting snarled in remark-highlight.js with
MDX v2. As a workaround, we can upgrade to the latest rehype version of
this plugin by using a wrapper function. I also updated/migrated
the slug plugin too. The remark-gfm plugin is unable to be updated using
this method currently.