Update links to the `connectors` repository (#327)
The [connectors](https://github.com/delta-io/connectors) repository has
been recently archived (on Jul 13, 2023). Its contents have been moved
into the [delta](https://github.com/delta-io/delta) repository, under
`connectors`.

Therefore, any link pointing to the archived repository should be
updated to point to the new location instead.

We update links from published blog posts as well, as that's probably
the right thing to do.

But we leave untouched any link that refers to an issue, a release, a
pull request or a branch, as those should still point to the archived
repository. Only links that refer to files or directories within the
repository are adjusted.

Co-authored-by: Denny Lee <denny.g.lee@gmail.com>